+++ packages/to-json-schema/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the library will be documented in this file.
 
 ## v1.0.0 (Month DD, YYYY)
 
+- Add support for `undefinedable` schema
 - Change Valibot peer dependency to v1.0.0
 - Change extraction of default value from `nullable`, `nullish` and `optional` schema